requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# HTTP & Scraping
requests>=2.31.0
httpx>=0.27.0  # Async HTTP client for Ollama API
beautifulsoup4>=4.12.0
lxml>=4.9.0
curl_cffi>=0.6.0  # JavDB compatibility

# Data Validation & Typing
pydantic>=2.0.0

# WebSocket
websockets>=12.0

# Utilities
pillow>=10.0.0
huggingface_hub>=0.20.0
onnxruntime>=1.18.0
numpy>=1.24.0

# Desktop GUI (Windows)
pywebview>=6.0
